+++ manifest.json
@@ -19,27 +19,27 @@
   "content_scripts": [
     {
       "matches": [
-        "*://www.parkrun.com.au/*/results/latestresults/",
-        "*://www.parkrun.co.at/*/results/latestresults/",
-        "*://www.parkrun.ca/*/results/latestresults/",
-        "*://www.parkrun.dk/*/results/latestresults/",
-        "*://www.parkrun.fi/*/results/latestresults/",
-        "*://www.parkrun.fr/*/results/latestresults/",
-        "*://www.parkrun.com.de/*/results/latestresults/",
-        "*://www.parkrun.ie/*/results/latestresults/",
-        "*://www.parkrun.it/*/results/latestresults/",
-        "*://www.parkrun.jp/*/results/latestresults/",
-        "*://www.parkrun.lt/*/results/latestresults/",
-        "*://www.parkrun.my/*/results/latestresults/",
-        "*://www.parkrun.co.nl/*/results/latestresults/",
-        "*://www.parkrun.co.nz/*/results/latestresults/",
-        "*://www.parkrun.no/*/results/latestresults/",
-        "*://www.parkrun.pl/*/results/latestresults/",
-        "*://www.parkrun.sg/*/results/latestresults/",
-        "*://www.parkrun.co.za/*/results/latestresults/",
-        "*://www.parkrun.se/*/results/latestresults/",
-        "*://www.parkrun.org.uk/*/results/latestresults/",
-        "*://www.parkrun.us/*/results/latestresults/"
+        "*://www.parkrun.com.au/*/results/*",
+        "*://www.parkrun.co.at/*/results/*",
+        "*://www.parkrun.ca/*/results/*",
+        "*://www.parkrun.dk/*/results/*",
+        "*://www.parkrun.fi/*/results/*",
+        "*://www.parkrun.fr/*/results/*",
+        "*://www.parkrun.com.de/*/results/*",
+        "*://www.parkrun.ie/*/results/*",
+        "*://www.parkrun.it/*/results/*",
+        "*://www.parkrun.jp/*/results/*",
+        "*://www.parkrun.lt/*/results/*",
+        "*://www.parkrun.my/*/results/*",
+        "*://www.parkrun.co.nl/*/results/*",
+        "*://www.parkrun.co.nz/*/results/*",
+        "*://www.parkrun.no/*/results/*",
+        "*://www.parkrun.pl/*/results/*",
+        "*://www.parkrun.sg/*/results/*",
+        "*://www.parkrun.co.za/*/results/*",
+        "*://www.parkrun.se/*/results/*",
+        "*://www.parkrun.org.uk/*/results/*",
+        "*://www.parkrun.us/*/results/*"
       ],
       "js": [
         "./dist/eventuate.js"
